Add icon rendering functions
